Makes Refresh use buildwindow instead
The code at line 117 in DlogEditorWindow.cs essentially did the same as BuildWindow with omission of setting windowEvents and calling Refresh. Although calling Refresh might not be as necessary the omission of windowEvents being set causes the Save Graph to stop function when you enter play mode and view the window again.